⚠️ Internal Docs ⚠️ Not Public API 👉 Official Docs Here

Module pyo3_pytests::othermod

https://github.com/PyO3/pyo3/issues/233

The code below just tries to use the most important code generation paths

Modules

double

Structs

ModClass

Statics

__PYO3_PYMODULE_DEF_OTHERMOD

Functions

PyInit_othermod⚠

This autogenerated function is called by the python interpreter when importing the module.

__pyfunction_double⚠

double

othermod